Move models to separate namespace
