+++ osu.Game/IPC/DataSources/PlayerStateWebSocketDataSource.cs
@@ -12,6 +12,7 @@ using osu.Game.Beatmaps;
 using osu.Game.Configuration;
 using osu.Game.Extensions;
 using osu.Game.IPC.Messages;
+using osu.Game.IPC.Models;
 using osu.Game.Online.Multiplayer;
 using osu.Game.Rulesets;
 using osu.Game.Rulesets.Mods;
