+++ osu.Game/IPC/Models/WebSocketMod.cs
@@ -4,7 +4,7 @@
 using System.Collections.Generic;
 using System.Text.Json.Serialization;
 
-namespace osu.Game.IPC.Messages
+namespace osu.Game.IPC.Models
 {
     public class WebSocketMod
     {
